Move dearchive command to domain layer
