+++ lib/pod/commands/dearchive/output.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module Pod
-  module Outputs
-    module Text
-      class Dearchive < ::Pod::Commands::BaseOutput
+  module Commands
+    module Dearchive
+      class Output < ::Pod::Commands::BaseOutput
         def call
           case @context[:details]
           when :episode_dearchived
